refactor: rename subdir to deep_stack
Remove DeepStackModule module and move submodules into DeepStack class
