+++ lib/deep_stack/face.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module DeepStackModule
+class DeepStack
   # APIs related to face recognition
   module Face
     #
