+++ lib/deep_stack/scene.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module DeepStackModule
+class DeepStack
   # Scene Recognition
   module Scene
     #
