+++ lib/deep_stack/version.rb
